# web/src/lib/api-client/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/users/": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create User",
                "operationId": "create_user_users__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCreateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{ethereum_address}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get User",
                "operationId": "get_user_users__ethereum_address__get",
                "parameters": [
                    {
                        "name": "ethereum_address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ethereum Address"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGetResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/credits/{user_id}": {
            "get": {
                "tags": [
                    "credits"
                ],
                "summary": "Get Credit Balance",
                "operationId": "get_credit_balance_credits__user_id__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreditBalanceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreditBalanceResponse": {
                "properties": {
                    "balance": {
                        "type": "number",
                        "title": "Balance"
                    },
                    "wallet_address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Wallet Address"
                    }
                },
                "type": "object",
                "required": [
                    "balance",
                    "wallet_address"
                ],
                "title": "CreditBalanceResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "UserCreate": {
                "properties": {
                    "ethereum_address": {
                        "type": "string",
                        "title": "Ethereum Address"
                    }
                },
                "type": "object",
                "required": [
                    "ethereum_address"
                ],
                "title": "UserCreate"
            },
            "UserCreateResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "ethereum_address": {
                        "type": "string",
                        "title": "Ethereum Address"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "ethereum_address"
                ],
                "title": "UserCreateResponse"
            },
            "UserGetResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "ethereum_address": {
                        "type": "string",
                        "title": "Ethereum Address"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "ethereum_address"
                ],
                "title": "UserGetResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}